fetch_and_store_stock_data/requirements.txt
google-cloud-bigquery
yfinance
requests
functions-framework
